Moved attributes under folder
