+++ src/Syringe.Web/Controllers/JsonController.cs
@@ -4,6 +4,7 @@ using Syringe.Core.Security;
 using Syringe.Core.Services;
 using Syringe.Core.Tasks;
 using Syringe.Core.Tests;
+using Syringe.Web.Controllers.Attribute;
 
 namespace Syringe.Web.Controllers
 {
